php-basercms: Import php-basercms-4.5.6 as wip/php-basercms
baserCMS is a japanese CMS (Content Management System) that offers
intuitive operation, high maintainability, and the freedom to
customize websites. It is open source software made by japanese people
for japanese people, by everyone. It is available free of charge, can
run on a variety of servers, and is easy to install.