opa: Update to 0.39.0
Changes:
0.39.0
======
Disk Storage
------------
The on-disk storage backend has been fully integrated with the OPA
server, and can now be enabled via configuration:

```
storage:
  disk:
    directory: /var/opa # put data here
    auto_create: true   # create directory if it doesn't exist
    partitions:         # partitioning is important for data storage,
      - /users/*        # please see the documentation
```

It is intended to enable the use of OPA in scenarios where the data
needed for policy evaluation exceeds the available memory.

The on-disk contents will persist among restarts, but should not
be used as a single source of truth: there are no backup mechanisms,
and certain data partitioning changes will require a start-over.
These are things that may get improved in the future.

For all the details, please refer to the configuration and detailled
Disk Storage section of the documentations.

0.38.1
=======
This is a bug fix release that addresses one issue when using opa
test with the --bundle (-b) flag, and a policy that uses the every
keyword.

0.38.0
=======
This release contains a number of fixes and enhancements.

It contains one backwards-incompatible change to the JSON representation
of metrics in Status API payloads, please see the section below.

Rich Metadata
-------------
It is now possible to annotate Rego policies in a way that can be
processed programmatically, using Rich Metadata.

```
allow {
  ...
  }
```

The available keys are:

 - title
 - description
 - authors
 - organizations
 - related_resources
 - schemas
 - scope
 - custom

Custom annotations can be used to annotate rules, packages, and
documents with whatever you specifically need, beyond the generic
keywords.

Annotations can be retrieved using the Golang library
or via the CLI, opa inspect -a.

All the details can be found in the documentation on Annotations.

Every Keyword
-------------
A new keyword for explicit iteration is added to Rego: every.

It comes in two forms, iterating values, or keys and values, of a
collection, and asserting that the body evaluates successfully for
each binding of key and value to the collection's elements:

```
every k, v in {"foo": "FOO", "bar": "BAR" } {
  upper(k) == v
}
```

To use it, import future.keywords.every or future.keywords.

For further information, please refer to the Every Keyword docs
and the new section on FOR SOME and FOR ALL in the Intro docs.

Backwards incompatible changes
------------------------------
The JSON representation of the Status API's payloads -- both for
GET /v1/status responses and the metrics sent to a remote Status
API endpoint -- have changed

0.37.2
=======
This is a bugfix release addressing two bugs:

- A regression introduced in the formatter fix for CVE-2022-23628.
- Support indices for appending to an array, conforming to JSON Patch (RFC6902)
for patch bundles.

0.37.1
=======
This is a bug fix release that reverts the
github.com/prometheus/client_golang upgrade in v0.37.0. The upgrade
exposed an issue in the serialization of Go runtime metrics in the
Status API

0.37.0
=======
This release contains a number of fixes and enhancements.

Strict Mode
-----------
There have been numerous possible checks in the compiler that fall
into this category:

They would help avoid common mistakes; but Introducing them would
potentially break some uncommon, but legitimate use.  We've thus
far refrained from introducing them. Now, a new "strict mode" allows
you to opt-in to these checks, and we encourage you to do so!

With OPA 1.0, they will become the new default behaviour.

For more details, see the docs on Compiler Strict Mode.

Delta Bundles
-------------
Delta bundles provide a more efficient way to make data changes by
containing patches to data instead of snapshots.  Using them together
with HTTP Long Polling, you can propagate small changes to bundles
without waiting for polling delays.

See the documentation for more details.