Import QBE compiler backend as wip/qbe
QBE aims to be a pure C embeddable backend that provides 70% of the
performance of advanced compilers in 10% of the code. Its small size serves
both its aspirations of correctness and our ability to understand, fix, and
improve it. It also serves its users by providing trivial integration and
great flexibility.