warsow: Now it actually works
This work is LONG overdue.

Back in 2019, Leonid Bobrov <mazocomp@disroot.org> wrote the big patch to
add *BSD support to Warsow in the first place. I was able to confirm that
the game with his patches worked on NetBSD/amd64 at the time.

I kept and uploaded his source changes to my GitHub account, but I never
actually made a pkgsrc package for it... until now.