goyq: Import goyq-4.25.1
yq is a lightweight and portable command-line YAML, JSON and XML processor.

yq uses jq like syntax but works with yaml files as well as json and
xml. It doesn't yet support everything jq does - but it does support
the most common operations and functions, and more is being added
continuously.