Add open-simh (version 2022-05-31).
Subject: [simh] Announcing the Open SIMH project
Date: Fri, 3 Jun 2022 15:52:08 -0400
To: simh@groups.io

Announcing the Open SIMH project

SIMH is a framework and family of computer simulators, initiated by Bob
Supnik and continued with contributions (large and small) from many
others, with the primary goal of enabling the preservation of knowledge
contained in, and providing the ability to execute/experience,
old/historic software via simulation of the hardware on which it ran.
This goal has been successfully achieved and has for these years created
a diverse community of users and developers.

This has mapped to some core operational principles:

First, preserve the ability to run old/historically significant
software. This means functionally accurate, sometimes bug-compatible,
but not cycle-accurate, simulation.

Second, make it reasonably easy to add new simulators for other hardware
while leveraging common functions between the simulators.

Third, exploit the software nature of simulation and make SIMH
convenient for debugging a simulated system, by adding non-historical
features to the environment.

Fourth, make it convenient for users to explore old system environments,
with as close to historical interfaces, by mapping them to new features
that modern host operating systems provide.

Fifth, be inclusive of people and new technology. It's serious work, but
it should be fun.

Previously, we unfortunately never spent the time to codify how we would
deliver on these concepts. Rather, we have relied on an informal use of
traditional free and open-source principles.

Recently a situation has arisen that compromises some of these
principles and thus the entire status of the project, creating
consternation among many users and contributors.

For this reason, a number of us have stepped up to create a new
organizational structure, which we call "The Open SIMH Project", to be
the keeper and provide formal governance for the SIMH ecosystem going
forward.  While details of the structure and how it operates are likely
to be refined over time, what will not change is our commitment to
maintaining SIMH as a free and open-source project, licensed under an
MIT-style license as shown on the "simh" repository page.

It is our desire that all of the past users and contributors will come
to recognize that the new organizational structure is in the best
interests of the community at large and that they will join us in it.
However, this iproject as defined, is where we intend to contribute our
expertise and time going forward.  At this point, we have in place the
following, although we foresee other resources being added in the future
as we identify the need and execute against them:

A Github "organization" for the project at https://github.com/open-simh

A Git repository for the simulators themselves at
https://github.com/open-simh/simh

The license for the SIMH simulator code base, found in LICENSE.txt in
the top level of the "simh" repository.

The "SIMH related tools" in https://github.com/open-simh/simtools. This
is also licensed under MIT style or BSD style open source licenses
(which are comparable apart from some minor wording differences).

A "SIMH Steering Group" -- project maintainers and guides.

The conventional git style process is used for code contributions, via
pull request to the project repository. The Steering Group members have
approval authority; this list is likely to change and grow over time.

By formalizing the underlying structure, our operational principles and
guidance can best benefit the community. These are being developed and
formalized, with a plan to publish them soon.

We have used our best judgment in setting up this structure but are open
to discussion and consideration of other ideas, and to making
improvements. Many of us have been part of different projects and
understand that past mistakes are real. We have tried to learn from
these experiences and apply the collected wisdom appropriately. We
desire to hear from the community as we update and refine the operating
structure for the Open SIMH project.

We hope for your patience and look forward to your support as we work to
refine the organization and be able to provide this wonderful resource
for anyone to use as we continue to evolve the technology provided by
the SIMH system.

     The SIMH Steering Group
        Clem Cole
        Richard Cornwell
        Paul Koning
        Timothe Litt
        Seth Morabito
        Bob Supnik