cuelang: Import cuelang-0.4.3 as wip/cuelang
CUE is an open-source data validation language and inference engine with
its roots in logic programming. Although the language is not a
general-purpose programming language, it has many applications, such as
data validation, data templating, configuration, querying, code generation
and even scripting. The inference engine can be used to validate data in
code or to include it as part of a code generation pipeline.

A key thing that sets CUE apart from its peer languages is that it merges
types and values into a single concept. Whereas in most languages types and
values are strictly distinct, CUE orders them in a single hierarchy (a
lattice, to be precise). This is a very powerful concept that allows CUE to
do many fancy things. It also simplifies matters. For instance, there is no
need for generics and enums, sum types and null coalescing are all the same
thing.