tcl-expect: Restore SHLIB_VERSION handling as in ../lang/tcl-expect
The patch to add SHLIB_VERSION to config.status didn't apply because
upstream used newer configure (as one might expect in 2018 vs 2010)
and the code for config.status has been changed to be more
declarative.  This commit adds a hunk with the same semantics, even
though it looks very little like the ../lang/tcl-expect patch.  Then,
the use of @SHLIB_VERSION@ is restored.