scalapack: version 2.2.0, patch changes
This updates to upstream 2.2.0 and as a notable local change does not hack
both shared and static libs into the build anymore. This is an unloved
endeavour and if we insist on static libs, we'd need to builds. But do we?

Tests are enabled now.

Upstream changes since 2.1.0:

New features:

    Allow compilation in ILP64 mode, PR #19

Major bug fixes:

    Use pointer macro to prevent integer overflow, PR #16

    Fix argument mismatches for implicitly defined functions (for gcc-10+ compatibility), PR #26

    No memory leaks on blacs_exit, PR #37