blueprint-compiler: add new package
GtkBuilder XML format is quite verbose, and many app developers
don't like using WYSIWYG editors for creating UIs. Blueprint files
are intended to be a concise, easy-to-read format that makes it
easier to create and edit GTK UIs.

Internally, it compiles to GtkBuilder XML as part of an app's build
system. It adds no new features, just makes the features that exist
more accessible.

Another goal is to have excellent developer tooling--including a
language server--so that less knowledge of the format is required.
Hopefully this will increase adoption of cool advanced features
like GtkExpression.