openmpi: simplify, config file handling, fixup
We do not go for concurrently installed MPI libraries, so no
MPI_PREFIX. The configuration file handling does not need
patches to the build.

OpenSHMEM is not that relevant for general use and actually
surpassed in the current MPI standard, so we try to get by without
it.

Scheduler integration options exist, but are not tested. It should
be trivial changes to PLIST, if at all, and handling of the
actual scheduler inside or outside of pkgsrc. On my HPC setups, the
scheduler comes as part of the base system or my own installation
before pkgsrc, just like MPI, in fact. MPI installed from pkgsrc is
useful for workstations that parallel applications are developed on.

Turning off Fortran (f09 option) has been tested, but other packages
(will) assume it is on, hence the default.

It is not the prettiest package, there are some unknowns and nits to
pick (like not using external hwloc), but it should all in all be
something useful in contrast to the ancient state in main.

Pkglint is happy, at least.