py-graphtage: new package
TODO: Nothing. Finished, import it if you find it useful.

Graphtage is a command-line utility and underlying library for
semantically comparing and merging tree-like structures, such as
JSON, XML, HTML, YAML, plist, and CSS files. Its name is a portmanteau
of "graph" and "graftage"-the latter being the horticultural practice
of joining two trees together such that they grow as one.