helm: Update to 3.9.4
Changes:
Helm 3.9.4
----------
Helm v3.9.4 is a security (patch) release. Users are strongly recommended to
update to this release.

While fuzz testing Helm, provided by the CNCF, a possible out of memory panic
was discovered with the strvals package. Out of memory panics cannot be
recovered from in Go. This can potentially be used to produce a denial of
service (DOS). More details are available in the advisory.

Helm 3.9.3
----------
Helm v3.9.3 is a patch release. Users are encouraged to upgrade for the best
experience. Users are encouraged to upgrade for the best experience.

Helm 3.9.2
----------
Helm v3.9.2 is a patch release. This release is the same Helm source as v3.9.1.
It is rebuilt on Go 1.17.12 which has numerous CVE fixes compared to the
version of Go used to build Helm v3.9.1.

Helm 3.9.1
----------
Helm v3.9.1 is a patch release. Users are encouraged to upgrade for the best
experience. Users are encouraged to upgrade for the best experience.

Helm 3.9.0
----------
Helm v3.9.0 is a feature release. This release, we focused on . Users are
encouraged to upgrade for the best experience.

Helm 3.8.2
----------
Helm v3.8.2 is a patch release. Users are encouraged to upgrade for the best
experience.

Helm 3.8.1
----------
Helm v3.8.1 is a patch release. Users are encouraged to upgrade for the best
experience.

Helm 3.8.0
----------
Helm v3.8.0 is a feature release. This release, we focused on OCI registries,
support for Kubernetes 1.23, and more. Users are encouraged to upgrade for the
best experience.

Notable Changes

  * OCI registry support for charts is now generally available. It has
    graduated out of being an experiment. Some changes over Helm 3.7 include:
      + When setting a chart in the dependencies as OCI the version can be set
        to a range, like other dependencies
      + SemVer tags that include build information can be pushed and used. OCI
        registries don't support + as a tag character. Helm translates the + to
        _ when stored as a tag
      + helm registry login now follows the same structure as the Docker CLI
        for storing credentials. The same location for registry configuration
        can be passed to both Helm and the Docker CLI
  * A --list-repo-url flag has been added to helm search hub to quickly show
    the repositories URL.
  * Release names are validated during installation
  * Added support for installing Helm on Windows via Cygwin
  * Added support for Kubernetes 1.23