roswell: Import roswell-21.10.14.111 as wip/roswell
Roswell is a Lisp implementation installer/manager, launcher, and
much more!

Roswell started out as a command-line tool with the aim to make
installing and managing Common Lisp implementations really simple
and easy.

Roswell has now evolved into a full-stack environment for Common
Lisp development, and has many features that makes it easy to test,
share, and distribute your Lisp applications.  With Roswell, we
aim to push the Common Lisp community to a whole new level of
productivity.