lwtools: update to 4.20.
2022-08-17: LWTOOLS 4.20. It is highly recommended that everyone using an older
version upgrade to this release. The big change is a fix to avoid relying on
undefined memory when deciding whether to register a symbol using the data or
code address. Also fix the basic output target to keep linesbelow 249
characters, a fix for numeric entry points in lwlink, and a couple other
miscelaneous fixes.

2022-01-25: LWTOOLS 4.19. The marquee change this time is fixing a macro
expansion bug related to positional parameters specified as {num}.
Additionally, lwlink will now not include any bytes for BSS sections when using
the raw target. If you were counting on the old behaviour, use the raw2 target.
Finally, this release includes a fix for the TFM compatibility instruction
sizes being accounted incorrectly.