cataclysm-dda-sdl: Import cataclysm-dda-sdl-0.F.3 as wip/cataclysm-dda-sdl
Cataclysm: Dark Days Ahead is a turn-based survival game set in a
post-apocalyptic world.

Struggle to survive in a harsh, persistent, procedurally generated world.

Scavenge the remnants of a dead civilization for food, equipment, or, if
you are lucky, a vehicle with a full tank of gas to get you the hell out of
Dodge.

Fight to defeat or escape from a wide variety of powerful monstrosities,
from zombies to giant insects to killer robots and things far stranger and
deadlier, and against the others like yourself, that want what you have...

This package contains the graphical SDL2-based interface with tiles.