qt6_qtmultimedia: fix build when gst-plugins1-bad is installed
Fix build when gst-plugins1-bad is installed, cmake finds
gstreamer photography using find_path and find_library that
looks outside of the buildlink system, the build then breaks
when the compiler  can't find the headers for photography.
Add buildlink that includes the photography module so that
the build always succeeds.