amnesia-tdd: Apply patches to make it actually playable
Now it actually plays on NetBSD/amd64 9.3, haha.

My previous patch which added NetBSD support (already integrated upstream)
was not complete -- it allowed the code to build with pkgsrc, but I didn't
have a copy of the game to properly test it. Until now.

So I'm submitting this other patch in the meanwhile.