xtrlock: add support for systems using a shadow file.
This adds support for Linux, SunOS, HP-UX and other systems relying on
/etc/shadow to store the password database, by including shadow.h and
using getspnam() instead of getpwnam().

If a `shadow' group is present, the xtrlock binary is installed sgid
"shadow". The group name can be changed to whatever group grants read
permissions to the password file, by explicitly setting the SHADOWGRP
variable.