xtrlock: removing TODO - The package has been succesfully built and tested on NetBSD, SunOS and Linux. A user reported it's working on OpenBSD too.
-
The SUID/SGID bit is required to retrieve the password/shadow entry, as
it's the case of x11/slock [1].

A workaround is very unlikely, unless the program is patched to use PAM,
which would imply rewriting large porptions of xtrlock  and maintaining
a downstream fork. A proof of concept is provided by this (old, and
unfortunately abandoned) fork [2].

This effort however looks beyond the scope of pkgsrc. Alternatives not
requiring suid/sgid are already available in repo.

[1] https://git.suckless.org/slock/file/slock.c.html
[2] https://github.com/aanatoly/xtrlock-pam/blob/master/src/xtrlock.c