doomlegacy-devel: Update to SVN revision 1636
- commit revision 1636
OpenGL display, wall textures would distort when a fireball or rocket
would pass by.
This would happen on any wall under a 3d floor (but not the top floor).
Hitting the top wall with a plasma, would cause distortion of the
lower walls.
Disabling dynamic lights would stop the distortion.

This occured because the HWR_SplitWall function was going through the
light planes, displaying each, using the GL vertexes as if they would
not change.
The dynamic lights overwrites the tow and sow of the GL vertexes with
light values.  As the x,y,z values were the same, and as the dynamic
lights are projected after the wall is projected, for most walls this
was more efficient.
Fixed the HWR_SplitWall function by adding explicit saving of the sow
and tow values, restoring them before each call to project the wall.

Also added some conditional code that would give the dynamic light
projection its own copy of the GL vertexes.  This also worked, but
was much more expensive.

- commit revision 1635
 Optimize some sprite clipping loops.