rust: checkpoint not-yet-finished support for mipsel-unknown-netbsd.
For some yet-inexplicable reason, LLVM detects that it needs
to use -latomic for 64-bit atomics, but when it comes to linking
the mipsel target executables, that knowledge appears to be gone
(or wasn't transferred to the rust build itself) and we end up with

          threads_pthread.c:(.text.CRYPTO_atomic_add+0x10): undefined reference to `__atomic_is_lock_free'

which is during linking of

work/rustc-1.64.0-src/build/x86_64-unknown-netbsd/stage1-tools/mipsel-unknown-netbsd/release/deps/cargo-e77c167e2db88b27

The above symbol is present in -latomic, but if the gcc wrapper
invocation doesn't mention the library, the above is the result.
Assistance / hints appreciated.

The target root in /u/mipsel/dest is a "plain" 9.3 mipsel tree,
but with the libatomic package installed using "pkg_add -m mipsel
-P /u/mipsel/dest ...", and the binary can be had from

ftp://ftp.netbsd.org//pub/NetBSD/misc/he/libatomic-mipsel-9.3-12.2.0.tgz

The 9.3 mipsel tools should be placed in /u/mipsel/tools, all ref.
the cross.mk Makefile fragment.