plan9port:  plan9/bin/auxstats needs sgid bit.
auxstats needs read access to /dev/kmem on non-Linux systems.
BSDs provide a kmem group for this, so we'll use that.
On SunOS this is not required, as /dev/kmem can be read be the standard
user.
See https://9fans.github.io/plan9port/man/man1/stats.html

Makefile slightly reworked to be more readable and easy to follow.