py-qt5: Fully depend on py-sip-qt5
It turns out that py-sip-qt5 provides qt5 bindings that have been
created with sip, not support for using sip with qt5.  It is not
possible to use py-qt5 without those files also being installed.
Therefore, change TOOL_DEPENDS to DEPENDS.

(It is possible DESCR of both should be updated, and that programs that
only want to use qt5 (rather than also run qt5 build tools) should
simply use py-sip-qt5 instead of py-qt5, but that remains to be
understood.)