prometheus-lts: Import prometheus-2.37.1nb3
The Prometheus monitoring system and time series database.

Prometheus, a Cloud Native Computing Foundation project, is a systems and
service monitoring system. It collects metrics from configured targets at
given intervals, evaluates rule expressions, displays the results, and
can trigger alerts if some condition is observed to be true.

Intended to be used to track the LTS version of Prometheus.