rust: fix our "mipsel" target so that it can work.
Earlier failed attempts:
 * cpu = "mips1", would have been more "portable" to older CPUs,
   but llvm failed to generate code for this cpu in at least one
   instance.  The "mips1" target in llvm is marked "highly experimental",
   so this is perhaps not surprising.
 * cpu = "mips3", an earlier mistake.  Produces "of course" 64-bit
   instructions, so gives "illegal instruction" on e.g. cobalt and
   32-bit-only CPUs.  Despite this also being marked "highly experimental"
   in llvm, this build completed.

So ... we give up some portability to the older CPUs, and generate
code for cpu = "mips32" with soft-float, and this produces a working
rustc / cargo combination, at least to the point that devel/cbindgen
can be built using a MIPSSIM kernel on qemu.

A native build of rust has yet to be attempted, it will be difficult
with the 508MB max memory limit for the 32-bit mips port, so the
next best is to try in a o32 chroot on a mips64 kernel which can
now get more memory.