perl5-devel: update to 5.37.8
The use of C<(?{ ... })> and C<(??{ ... })> in a pattern disables various
optimisations globally in that pattern. This may or may not be desired by the
programmer. This release adds the C<(*{ ... })> and C<(**{ ... })>
equivalents. The only difference is that they do not and will never disable
any optimisations in the regex engine. This may make them more unstable in the
sense that they may be called more or less times in the future, however the
number of times they execute will truly match how the regex engine functions.
For example, certain types of optimisation are disabled when C<(?{ ... })> is
included in a pattern, so that patterns which are O(N) in normal use become
O(N*N) with a C<(?{ ... })> pattern in them. Switching to C<(*{ ... })> means
the pattern will stay O(N).

 - L<Config::Perl::V> has been upgraded from version 0.34 to 0.35.
 - L<ExtUtils::Install> has been upgraded from version 2.20 to 2.22.
 - L<ExtUtils::MakeMaker> has been upgraded from version 7.64 to 7.66.
 - L<feature> has been upgraded from version 1.78 to 1.79.
 - L<File::Find> has been upgraded from version 1.41 to 1.42.
 - L<IO::Zlib> has been upgraded from version 1.11 to 1.14.
 - L<JSON::PP> has been upgraded from version 4.12 to 4.16.
 - L<Math::Complex> has been upgraded from version 1.5903 to 1.6.
 - L<Math::Trig> has been upgraded from version 1.23_01
 - L<Module::CoreList> has been upgraded from version 5.20221220 to 5.20230120.
 - L<podlators> has been upgraded from version 5.00 to 5.01.
 - L<POSIX> has been upgraded from version 2.10 to 2.11.
 - L<re> has been upgraded from version 0.43 to 0.44.
 - L<Storable> has been upgraded from version 3.28 to 3.29.
 - L<threads> has been upgraded from version 2.33 to 2.34.
 - L<XS::APItest> has been upgraded from version 1.28 to 1.30.