email-oauth2-proxy: add version 2022.12.14.
Transparently add OAuth 2.0 support to IMAP/POP/SMTP client
applications, scripts or any other email use-cases that don't
support this authentication method.

Motivation and capabilities

Email services that support IMAP, POP and/or SMTP access are
increasingly requiring the use of OAuth 2.0 to authenticate
connections, but not all clients support this method. This script
creates a simple local proxy that intercepts the traditional
IMAP/POP/SMTP authentication commands and transparently replaces
them with the appropriate SASL (X)OAuth 2.0 commands and credentials.
Your email client can continue to use the login or auth/authenticate
options, with no need to make it aware of OAuth's existence. The
proxy works in the background with a menu bar/taskbar helper or as
a system service, and is compatible with macOS, Windows and Linux.

TODO:
  testing...