py-cloud-init: add version 22.4.2.
Program to consume & act on cloud initialization data.

This one does not build or install on NetBSD, there is
a fair bit of GNUisms and Linux-isms in the self-tests,
which Makefile insists on must pass, the code itself may
also have those.

Also need to figure out how this package is supposed to be built
and installed, instructions on those points appear to not be included
in the source distribution, at least not prominently visible.

The package dependency graph for this one is rather large,
it includes both python (of course), perl5, and rust(!).  Some of
those might not be run-time dependencies, though, but still...

See the TODO file for "unfinished business", including a more
detailed review of the failing self-tests.

The lighter-weight approach of the NetBSD in-tree script at

  src/distrib/utils/embedded/files/ec2_init

might warrant a closer look as a (much!) simpler and light-weight
alternative to this to do the absolute minimum.